requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
pillow>=9.4.0

# Data processing
opencv-python>=4.7.0
mediapipe>=0.10.0
scikit-image>=0.20.0
h5py>=3.8.0
tqdm>=4.65.0
pytube>=12.1.0
gdown>=4.7.0

# Model export & optimization
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0
tensorrt>=8.5.0

# API & Web Services
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
websockets>=11.0.0
httpx>=0.24.0
pydantic>=1.10.0

# Dashboard
streamlit>=1.22.0
plotly>=5.14.0
altair>=5.0.0
watchdog>=3.0.0

# Enterprise Integrations
# Authentication
pyjwt>=2.6.0
passlib>=1.7.4
python-jose>=3.3.0

# Metrics
prometheus-client>=0.16.0
psutil>=5.9.0

# Kafka
confluent-kafka>=2.0.0

# MQTT
paho-mqtt>=2.1.0

# Webhooks
aiohttp>=3.8.0

# DevOps & Utilities
pytest>=7.3.0
ruff>=0.0.262
black>=23.3.0
isort>=5.12.0
pre-commit>=3.3.0
python-dotenv>=1.0.0

# Optimization
optuna>=3.1.0

# Development dependencies
-r requirements.txt

# Testing
pytest-cov>=4.1.0
pytest-mock>=3.10.0
coverage>=7.2.0

# Additional code quality tools
mypy>=1.3.0
bandit>=1.7.0

# Documentation
sphinx>=6.1.0
sphinx-rtd-theme>=1.2.0
sphinx-autoapi>=2.1.0

# Development tools
ipython>=8.13.0
jupyter>=1.0.0
notebook>=6.5.0

# Performance analysis
memory-profiler>=0.61.0
snakeviz>=2.2.0

# Security testing
safety>=2.3.0
